WIP: Step 1 for issue #8.
Refactored out injection code from the plugin class.  The web view
specific code will be in delegates specific to the view type which is
detected by the plugin.  This commit establishes the base and gets
injection working for the UIWebView.

Next step is to create the delegate for WKWebView and verify the
approach will work for it.  After that I'll work the retry logic in.